+++ src/test/java/playground/domain/document/ApprovalStateTest.java
@@ -1,4 +1,4 @@
-package learning;
+package playground.domain.document;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
