+++ src/test/java/playground/domain/documentapproval/DocumentApprovalTest.java
@@ -1,4 +1,4 @@
-package learning;
+package playground.domain.documentapproval;
 
 import fixture.DocumentApprovalFixture;
 import fixture.UserFixture;
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
+import playground.domain.document.ApprovalState;
+import playground.domain.user.User;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
